[T] general functionality put to separate header
